Create Query_JOIN_Top_Countries
I reference the ERD to determine which tables I need to pull data from (customer and country).  Using multiple joins. Customer → address → city → country. 
In the query, I included country as the top 10 countries and COUNT(a.customer_id).  I chose INNER JOIN because I only needed a small amount of data that was present in the two tables. 
After the join statements, I grouped by d.country and ordered the number of count of customers from highest to lowest, and set a limit to 10 since we only need the top 10 countries where they have the majority of their customers.